pyproject.toml
[project]
name = "cheatsheet-gen"
version = "0.1.0"
description = "Generates cheat sheets from pptx or pdf files"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imgkit>=1.2.3",
    "markdown>=3.7",
    "markdown-strings>=3.4.0",
    "pdf2image>=1.17.0",
    "python-pptx>=1.0.2",
    "ollama>=0.3.3",
]
